+++ boards/PDM/Test/Inc/Test_Pdm.h
@@ -3,10 +3,4 @@
 #include <fff.h>
 #include <gtest/gtest.h>
 
-template <typename T>
-void TearDownObject(T *&object, void (*destroy_func)(T *object))
-{
-    ASSERT_TRUE(object != NULL);
-    destroy_func(object);
-    object = NULL;
-}
+#include "Test_Utils.h"
