+++ boards/shared/Test_Utils/Test_Utils.h
@@ -1,0 +1,9 @@
+#pragma once
+
+template <typename T>
+void TearDownObject(T *&object, void (*destroy_func)(T *object))
+{
+ASSERT_TRUE(object != NULL);
+destroy_func(object);
+object = NULL;
+}
